Expose factory not protocol.
