+++ worker/buildbot_worker/pb.py
@@ -29,6 +29,7 @@ from buildbot_worker.base import BotBase
 from buildbot_worker.base import WorkerBase
 from buildbot_worker.base import WorkerForBuilderBase
 from buildbot_worker.pbutil import ReconnectingPBClientFactory
+from buildbot_worker.util import HangCheckFactory
 
 
 class UnknownCommand(pb.Error):
@@ -178,8 +179,13 @@ class Worker(WorkerBase, service.MultiService):
         bf = self.bf = BotFactory(buildmaster_host, port, keepalive, maxdelay)
         bf.startLogin(
             credentials.UsernamePassword(name, passwd), client=self.bot)
-        self.connection = c = internet.TCPClient(buildmaster_host, port, bf)
+        self.connection = c = internet.TCPClient(
+            buildmaster_host, port,
+            HangCheckFactory(bf, hung_callback=self._hung_connection))
         c.setServiceParent(self)
+
+    def _hung_connection(self):
+        log.msg("connection attempt timed out (is the port number correct?)")
 
     def startService(self):
         WorkerBase.startService(self)
