+++ worker/buildbot_worker/util/__init__.py
@@ -19,14 +19,14 @@ import itertools
 import textwrap
 import time
 
-from ._hangcheck import HangCheckProtocol
+from ._hangcheck import HangCheckFactory
 
 __all__ = [
     "remove_userpassword",
     "now",
     "Obfuscated",
     "rewrap",
-    "HangCheckProtocol",
+    "HangCheckFactory",
 ]
 
 def remove_userpassword(url):
